requirements.txt
APScheduler==3.10.4
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
oauthlib==3.2.2
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==1.3.1
six==1.16.0
tweepy==4.14.0
tzlocal==5.2
urllib3==2.2.3
